Add support for entity interfaces (#5)
