+++ aes-processor/src/main/java/com/ijioio/aes/annotation/processor/exception/CodeGenerationException.java
@@ -1,0 +1,15 @@
+package com.ijioio.aes.annotation.processor.exception;
+
+public class CodeGenerationException extends ProcessorException {
+
+	private static final long serialVersionUID = 1643374694628890189L;
+
+	/**
+	 * Constructs a new exception with the specified details.
+	 *
+	 * @param message with the details of the exception
+	 */
+	public CodeGenerationException(String message) {
+		super(message);
+	}
+}
