+++ aes-sandbox/src/main/java/com/ijioio/aes/sandbox/Number.java
@@ -1,0 +1,6 @@
+package com.ijioio.aes.sandbox;
+
+public interface Number {
+
+	public String getNumber();
+}
